+++ modules/custom/az_core/az_core.services.yml
@@ -20,25 +20,50 @@ services:
       - '@config_sync.snapshotter'
       - '@config_update.config_list'
       - '@module_handler'
-  cache.az_migration_http_cache:
+  cache.az_core_http_cache:
     class: Drupal\Core\Cache\CacheBackendInterface
     tags:
       - { name: cache.bin }
-    factory: cache_factory:get
-    arguments: [az_migration_http_cache]
-  az_core.az_migration_http_cache_middleware:
-    class: Drupal\az_core\AZCoreGuzzleCache
-    arguments: ['@config.factory','@cache.az_migration_http_cache', 600]
+    factory: ['@cache_factory', 'get']
+    arguments:
+      - az_core_http_cache
+  cache.az_core_http_cache_memory:
+    class: Drupal\Core\Cache\MemoryCache\MemoryCacheInterface
+    tags:
+      - { name: cache.bin.memory, default_backend: cache.backend.memory.memory }
+    factory: ['@cache_factory', 'get']
+    arguments:
+      - az_core_http_cache_memory
+  cache.az_core_http_cache_chained:
+    class: Drupal\Core\Cache\BackendChain
+    calls:
+      - [appendBackend, ['@cache.az_core_http_cache_memory']]
+      - [appendBackend, ['@cache.az_core_http_cache']]
+    tags: [{ name: cache.bin.memory }]
+  az_core.guzzlecache_storage:
+    class: Drupal\az_core\AZCoreGuzzleCacheStorage
+    arguments:
+      - '@cache.az_core_http_cache_chained'
+  az_core.guzzlecache_strategy_polite:
+    class: Drupal\az_core\Strategy\AZPoliteCacheStrategy
+    arguments:
+      - '@az_core.guzzlecache_storage'
+      - 600
+  az_core.az_migration_guzzlecache_middleware:
+    class: Drupal\az_core\AZCoreGuzzleCacheMiddleware
+    arguments: ['@config.factory','@az_core.guzzlecache_strategy_polite']
     tags:
       - { name: az_core_migration_http_client_middleware, priority: 0 }
   az_core.az_migration_http_retry_middleware:
     class: Drupal\az_core\AZCoreRetryMiddleware
-    arguments: [5]
+    arguments:
+      - 5
     tags:
       - { name: az_core_migration_http_client_middleware, priority: 100 }
   az_core.az_migration_http_handler_stack_configurator:
     class: Drupal\Core\Http\HandlerStackConfigurator
-    arguments: ['@service_container']
+    arguments:
+      - '@service_container'
     tags:
       - { name: service_id_collector, tag: az_core_migration_http_client_middleware }
   az_core.az_migration_http_handler_stack:
@@ -47,7 +72,8 @@ services:
     configurator: ['@az_core.az_migration_http_handler_stack_configurator', configure]
   az_core.az_migration_http_client_factory:
     class: Drupal\Core\Http\ClientFactory
-    arguments: ['@az_core.az_migration_http_handler_stack']
+    arguments:
+      - '@az_core.az_migration_http_handler_stack'
   az_core.az_migration_http_client:
     class: GuzzleHttp\Client
     factory: ['@az_core.az_migration_http_client_factory', 'fromOptions']
